Move stuff into Pipeline/Middleware namespace
